Refactor tests and add Ruby test
