+++ test/integration/configs/instrumenter-config-ruby.yml
@@ -1,0 +1,6 @@
+routes:
+  patterns:
+    - /users/:user_id
+  unmatch: path
+otel_metrics_export:
+  endpoint: http://otelcol:4318
